requirements.txt
asgiref==3.11.1
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
numpy==2.2.6
pgvector==0.4.2
psycopg2-binary==2.9.9
PyJWT==2.12.1
sqlparse==0.5.5
typing_extensions==4.15.0
